Extract `FeedLoader` tests helpers into a shared protocol extension to remove duplication
